pyproject.toml
[project]
name = "gatewayapi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "ariadne-codegen>=0.14.0",
    "booktype>=1.5",
    "configs>=3.0.3",
    "confluent-kafka==2.10.0",
    "django>=5.2.1",
    "django-filter==25.1",
    "djangorestframework==3.16.0",
    "graphene==3.4.3",
    "graphene-django==3.2.3",
    "graphql-core==3.2.6",
    "graphql-relay==3.2.0",
    "jwt>=1.3.1",
    "kafka-python>=2.2.10",
    "markdown==3.8",
    "phpserialize>=1.3",
    "promise==2.3",
    "python-dateutil==2.9.0.post0",
    "python-decouple==3.8",
    "python-dotenv>=1.1.0",
    "redis==6.2.0",
    "requests>=2.32.3",
    "sentry-sdk>=2.29.1",
    "six==1.17.0",
    "text-unidecode==1.3",
    "typing-extensions==4.13.2",
    "uvicorn>=0.34.3",
]

[tool.ariadne-codegen]
queries_path = "queries.graphql"
schema_path = "schema.graphql"
async_client = false
